requirements.txt
Flask==2.3.3
python-telegram-bot==20.7
sqlite3
gunicorn==21.2.0
numpy